divide Dashboard component into DashboardSidebar, DashboardLayout and DashboardPage components
